Try to work around the way Travis checks out code
[This page](http://about.travis-ci.org/docs/user/build-configuration/)
talks about customizing the way Travis builds your repository, and it
says that "With the exception of cloning project repository and changing
directory to it, all of the above steps can be tweaked". Since we can't
change where our repo is cloned to, we can make a symbolic link instead.

Tested on my machine (I moved my install to src/github.com/limetext/lime)
but not tested on Travis yet.